requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
groq==1.2.0
anthropic==0.96.0
python-dotenv==1.2.2
python-multipart==0.0.26
httpx==0.28.1